Basic TxTransferActionView and TxSettingsActionView (#791)
* Basic TxTransferActionView
* Cover settings changes
* Enable / Disable Modules
* value should be BigInteger instead of String
* Amount and Arrow Margins and Padding
* Hide txConfirmations when MultisigExecutionDetails missing
* Switching to staging client gateway
* Show token icons
* Use executor instead of last signer
* Advanced after executed
* Copy address to clipboard
* Copy tx hash to clipboard
* Also hide dividers when hiding executed/created or confirmations list
* zero byte data section is not expandable
* Show 0 bytes for custom tx
* Increased test coverage

+++ data/src/test/java/io/gnosis/data/adapters/DataMoshiTest.kt
@@ -43,7 +43,7 @@ class DataMoshiTest {
                     tokenAddress = "0xc778417E063141139Fce010982780140Aa0cD5Ab".asEthereumAddress()!!,
                     tokenName = "Wrapped Ether",
                     tokenSymbol = "WETH",
-                    value = "1000000000000000000"
+                    value = "1000000000000000000".toBigInteger()
                 )
             ),
             executionInfo = null
@@ -65,7 +65,7 @@ class DataMoshiTest {
                 type = GateTransactionType.Custom,
                 dataSize = 0,
                 to = "0x2134Bb3DE97813678daC21575E7A77a95079FC51".asEthereumAddress()!!,
-                value = "3140000000000000000"
+                value = "3140000000000000000".toBigInteger()
             ),
             executionInfo = ExecutionInfoDto(
                 nonce = 8.toBigInteger(),
